requirements.txt
requests>=2.30.0
pytest>=7.3.0
pydantic>=1.10.7
allure-pytest>=2.13.0